src/pygiggle/hackerrank/test/test_alphabet_rangoli.py
"""Tests for Alphabet Rangoli challenge from HackerRank."""

import pytest

from ..alphabet_rangoli import rangoli


RANGOLI3 = """\
----c----
--c-b-c--
c-b-a-b-c
--c-b-c--
----c----\
"""

RANGOLI5 = """\
--------e--------
------e-d-e------
----e-d-c-d-e----
--e-d-c-b-c-d-e--
e-d-c-b-a-b-c-d-e
--e-d-c-b-c-d-e--
----e-d-c-d-e----
------e-d-e------
--------e--------\
"""

RANGOLI10 = """\
------------------j------------------
----------------j-i-j----------------
--------------j-i-h-i-j--------------
------------j-i-h-g-h-i-j------------
----------j-i-h-g-f-g-h-i-j----------
--------j-i-h-g-f-e-f-g-h-i-j--------
------j-i-h-g-f-e-d-e-f-g-h-i-j------
----j-i-h-g-f-e-d-c-d-e-f-g-h-i-j----
--j-i-h-g-f-e-d-c-b-c-d-e-f-g-h-i-j--
j-i-h-g-f-e-d-c-b-a-b-c-d-e-f-g-h-i-j
--j-i-h-g-f-e-d-c-b-c-d-e-f-g-h-i-j--
----j-i-h-g-f-e-d-c-d-e-f-g-h-i-j----
------j-i-h-g-f-e-d-e-f-g-h-i-j------
--------j-i-h-g-f-e-f-g-h-i-j--------
----------j-i-h-g-f-g-h-i-j----------
------------j-i-h-g-h-i-j------------
--------------j-i-h-i-j--------------
----------------j-i-j----------------
------------------j------------------\
"""


@pytest.mark.parametrize("size,expected", [
    (3, RANGOLI3),
    (5, RANGOLI5),
    (10, RANGOLI10),
])
def test_rangoli(size, expected):
    """it returns a rangoli of specified size"""
    assert rangoli(size) == expected
